Module Std.Traces

Loaded traces.

This is a static container for the traces. Frontends and plugins can load traces and add it to the repository.

val to_list : unit -> trace list

to_list () returns a list of currently loaded traces.

val enum : unit -> trace Bap.Std.seq

enum () enumerates all currently loaded traces

val add : trace -> unit

add trace registers a trace into the repository

val remove : trace -> unit

remove trace removes a trace from the repository, do nothing if it wasn't registered.