Module Trace.Monitor

Monitor defines an error handling policy.

type t = monitor

val ignore_errors : t

ignore_errors filters good events and silently drops error events

warn_on_error on_error same as ignore_errors but calls on_error function when an error has occurred

val warn_on_error : (Core_kernel.Error.t -> unit) -> t

warn_on_error on_error same as ignore_errors but calls on_error function when an error has occurred

val fail_on_error : t

fail_on_error will fail with an error Error.raise error

val stop_on_error : t

stop_on_error will silently finish a stream in case of error.

val pack_errors : (Core_kernel.Error.t -> event) -> t

pack_errors pack will transform any occurred error into event using pack function.

val create : (Core_kernel.Error.t -> step) -> t

create filter creates a user defined monitor from function filter that is applied to a sequence of events or errors, and returns a sequence of events.